job-manager: refactor restart logic
Move the remaining portion of the restart logic out
of job-manager.c, into restart.c.  This aligns its
structure with other "sub modules" of the job manager.
