+++ src/modules/job-manager/job-manager.c
@@ -60,40 +60,6 @@ static void priority_cb (flux_t *h, flux_msg_handler_t *mh,
     priority_handle_request (h, ctx->queue, ctx->event_ctx, msg);
 }
 
-/* reload_map_f callback
- * This is called for each job encountered in the KVS during startup.
- * The 'job' struct is valid only during the callback.
- * queue_insert() increments the 'job' usecount upon successful insert.
- */
-static int restart_map_cb (struct job *job, void *arg)
-{
-    struct job_manager_ctx *ctx = arg;
-
-    if (queue_insert (ctx->queue, job, &job->queue_handle) < 0)
-        return -1;
-    if (job->state == FLUX_JOB_NEW)
-        job->state = FLUX_JOB_SCHED;
-    if (job->state == FLUX_JOB_SCHED) {
-        if (alloc_do_request (ctx->alloc_ctx, job) < 0) {
-            flux_log_error (ctx->h, "%s: error notifying scheduler of job %llu",
-                __FUNCTION__, (unsigned long long)job->id);
-        }
-    }
-    return 0;
-}
-
-/* Load any active jobs present in the KVS at startup.
- */
-static int restart_from_kvs (flux_t *h, struct job_manager_ctx *ctx)
-{
-    int count;
-
-    if ((count = restart_map (h, restart_map_cb, ctx)) < 0)
-        return -1;
-    flux_log (h, LOG_DEBUG, "%s: added %d jobs", __FUNCTION__, count);
-    return 0;
-}
-
 static const struct flux_msg_handler_spec htab[] = {
     { FLUX_MSGTYPE_REQUEST, "job-manager.submit", submit_cb, 0},
     { FLUX_MSGTYPE_REQUEST, "job-manager.list", list_cb, FLUX_ROLE_USER},
@@ -127,7 +93,7 @@ int mod_main (flux_t *h, int argc, char **argv)
         flux_log_error (h, "flux_msghandler_add");
         goto done;
     }
-    if (restart_from_kvs (h, &ctx) < 0) {
+    if (restart_from_kvs (h, ctx.queue, ctx.alloc_ctx) < 0) {
         flux_log_error (h, "restart_from_kvs");
         goto done;
     }
